backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.20

# AI Integration - OpenAI (using gpt-4o-mini model)
openai>=1.54.0

# IFC File Parsing
ifcopenshell==0.8.4

# Configuration & Data Validation
python-dotenv>=1.0.0
pydantic>=2.8.0
